migrated more code
[ # ] Moved Moderator commands from Commands.cs into StaffCommands.cs for better structure
[ # ] Fixed issue with Components trying to call non-existent function
[ + ] Added Database modeel for Ticket categories
[ # ]  Increased Ticket Reason from 1024 to 2048
[ # ] Commented out TicketCategory to migrate to the Database model object later, awaiting deletion
[ # ] Changed needed permissions for StaffCommands from Administrator to BanMembers so it applies to all moderators, important commands remain admin only and are overwritten as such in the command declarations
