+++ LuminateDiscordBot/Objects/TicketCategory.cs
@@ -9,16 +9,12 @@ namespace LuminateDiscordBot.Objects
 {
     public class TicketCategory
     {
+        /*
         public required string TicketTopic { get; set; }
         public required string CategoryAliasList { get; set; }
         public required string TicketDataName { get; set; }
         public required string TicketDataDescription { get; set; }
         public required string TicketDataAutoResponse { get; set; }
-
-        public List<string>? GetCategoryAliases()
-        {
-            return JsonSerializer.Deserialize<List<string>>(Convert.FromBase64String(this.CategoryAliasList));
-        }
-        
+        public required bool */
     }
 }
